requirements.txt
grpcio
grpcio-tools
streamlit
torch
diffusers
transformers
accelerate
Pillow
streamlit
xformers
sentencepiece
sacremoses
indic-transliteration
ctransformers
scikit-learn
vosk
sounddevice
openai-whisper
pyaudio
